+++ src/main/java/blusunrize/immersiveengineering/common/data/ItemModels.java
@@ -215,7 +215,7 @@ public class ItemModels extends LoadedModelProvider
 
 		addItemModels(
 				"tool_", mcLoc("item/handheld"), Tools.hammer, Tools.wirecutter, Tools.screwdriver,
-				Tools.manual, Tools.steelPick, Tools.steelShovel, Tools.steelAxe, Tools.steelSword
+				Tools.manual, Tools.steelPick, Tools.steelShovel, Tools.steelAxe, Tools.steelHoe, Tools.steelSword
 		);
 		addItemModels("", IEItems.Misc.wireCoils.values().toArray(new Item[0]));
 		addItemModels("", IEItems.Misc.graphiteElectrode);